Move all code into a single file.
Refactor tests to run with refactor
